requirements.txt
boto>=2.7
boto3>=1.4.4
bunch
decorator
logutils
pyyaml
requests
stevedore
simplejson
dill